edited asset file structure
